+++ oci/database_backup_destination_test.go
@@ -201,11 +201,13 @@ func TestDatabaseBackupDestinationResource_basic(t *testing.T) {
 			},
 			// verify resource import
 			{
-				Config:                  config,
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{},
-				ResourceName:            resourceName,
+				Config:            config,
+				ImportState:       true,
+				ImportStateVerify: true,
+				ImportStateVerifyIgnore: []string{
+					"mount_type_details",
+				},
+				ResourceName: resourceName,
 			},
 		},
 	})